requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
flake8==3.6.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
Pillow==5.3.0
pluggy==0.8.0
py==1.7.0
pycodestyle==2.4.0
pyflakes==2.0.0
pytest==4.0.2
six==1.12.0
Werkzeug==0.14.1
